server/requirements.txt
Django==1.10.5
django-cors-headers==2.0.2
django-filter==1.0.2
djangorestframework==3.5.3
psycopg2==2.6.2
